[v1.1.0-pre1] The real almost 1.20.6 update
Due to the changes with the Mixin system in new version of Minecraft and Fabric, many mods have been removed temporarily until they are patched for the new setup. I really just wanted to have a *version* out for people to use in the meantime. All removals are listed below, some which break OptiFine compatibility and remove a few performance features.

- Updated to MC 1.20.6 and Fabric 0.15.11
- The follwing were removed temporarily:
  - [Enhanced Block Entities](https://modrinth.com/mod/ebe)
  - [Language Reload](https://modrinth.com/mod/language-reload)
  - [CIT Resewn](https://modrinth.com/mod/cit-resewn)
  - [Inventory Profiles Next](https://modrinth.com/mod/inventory-profiles-next)